invoice-cli/pyproject.toml
[project]
name = "invoice-cli"
version = "0.1.0"
description = "Fetch and organize invoice emails from Gmail using AI classification"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [
    { name = "Desplega", email = "contact@desplega.ai" }
]
dependencies = [
    "typer>=0.15.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "baml-py>=0.76.0",
    "tomli-w>=1.0.0",
    "pdfplumber>=0.11.0",
    "textual>=6.11.0",
]

[project.scripts]
invoice-cli = "invoice_cli.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
